Split App into more discrete components
